+++ resources/js/components/Bible/LeafComponent.vue
@@ -1,0 +1,24 @@
+<template>
+    <div class="container">
+        <div class="row justify-content-center">
+            <div class="col-md-8">
+                <div class="card">
+                    <div class="card-header">Leaf Component</div>
+
+                    <div class="card-body">
+                        I'm a {{ type }} Leaf Component.
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+</template>
+
+<script>
+    export default {
+        mounted() {
+            console.log('Component mounted.')
+        },
+        props:['type']
+    }
+</script>