+++ resources/views/bible/index.blade.php
@@ -6,8 +6,10 @@
             {{-- <router-link to="/read/example">Go to Example</router-link> --}}
 
             {{-- <router-view></router-view> --}}
-            <index-component v-on:select-page="selectPage"></index-component>
+            <index-component :books="books" v-on:select-page="selectPage"></index-component>
             <page-component :page="selectedPage" :chapter-text="pageText"></page-component>
+            <leaf-component type="previous"></leaf-component>
+            <leaf-component type="next"></leaf-component>
         </div>
     </div>
 @endsection